after attack

+++ calculator.py
@@ -1,6 +1,7 @@
 
-a=int(input('enter first number '))
-b=int(input('enter second number'))
+a=float(input('enter first number '))
+b=float(input('enter second number'))
 sum=a+b
 print(sum)
+
